requirements.txt
aiogram==3.7.0
supabase==2.4.6
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
